Make MsSqlSpatial BoundingBox optional
RFC 7946 The GeoJSON Format states (https://tools.ietf.org/html/rfc7946):

> A GeoJSON object MAY have a member named "bbox" to include information on the coordinate range for its Geometries, Features, or FeatureCollections

There should be a way to make this property optional - this update provides an optional parameter to the method that does the conversion from SqlGeometry to GeoJSON so that this property can be made optional.

The default value is set to `true` so that it does not interfere with existing code that uses the library but provides an extension point for new development that may not want this property set.

Unit tests were added to verify that the behavior does not change when the interface is used instead of the concrete implementation.

Resolves: #18